requirements.txt
fastapi==0.111.0
mysql-connector-python==8.4.0
uvicorn==0.29.0
python-dotenv==1.0.1
holidays==0.47
pytz==2024.1
httpx==0.27.0
openai==1.35.13
python-dateutil==2.9.0.post00
# Dependencies for payment proof analyzer
PyPDF2==3.0.1
pdf2image==1.17.0
aiohttp==3.9.1
Pillow==10.1.0
playwright==1.44.0
pandas==2.0.3
openpyxl==3.1.3
xlrd==2.0.1

# Dependencies for Gmail API
google-api-python-client==2.134.0
google-auth-oauthlib==1.2.0
